Add multi-distro support to install-docker
This change adds the ability for the zuul provided role for docker
installation to support multiple distros automatically. With this
change the role will dynamically load variables based on the
target operating system.

While the default to install from upstream reamains unchanged, the
role will now fall back to installing using the distro packages
in the event there's is a fault with the upstream tasks or if an
upstream task file has not been implemented.

> stubbs for upstream implementation task files have been created.

With this change the role will now support, debian (and family),
opensuse, and red hat (and family). The upstream installation process
is using the documented steps from upstream docker, as found here
[0].

The default "ubuntu_gpg_key" as renamed to "docker_gpg_key" as its
now used for more than one distro.

The daemon.json.j2 template has been removed. This template was
stored here but unused by this role.

[0] https://docs.docker.com/install/linux/docker-ce/fedora/

Change-Id: Ie52444f8c62d6e2f86ce9ba1c17795dd9a6471d2
Signed-off-by: Kevin Carter <kecarter@redhat.com>