Prepare: Use PDK to build puppet module
The puppet module command is deprecated and
is replaced by using the PDK [1] [2].

The difference between the two commands is only
the output in the pkg directory.

The old puppet module command will create a folder and a tarball with the module.

$ls pkg/
openstack-openstacklib-14.1.0/
openstack-openstacklib-14.1.0.tar.gz

While the PDK will only create the tarball in the
pkg folder. This should be no problem since I can't
find anywhere were we are using the generated folder
in the pkg folder.

This is part 1 in a series of two changes, it contains a role to install
the package dependencies (run in pre.yaml before revoke-sudo is called),
the followup change adds the role to actually build. We need those two
steps to not introduce breakage since the users of build-puppet-module
need to be updated before we can change build-puppet-module.

Change-Id: Ief299751c001542b10fcdf44f1b40a169135877a
Needed-By: https://review.opendev.org/667589
Needed-By: https://review.opendev.org/667595
Co-Authored-By: Tobias Urdin <tobias.urdin@binero.se>