Use PDK to build puppet module
The puppet module command is deprecated and
is replaced by using the PDK [1] [2].

The difference between the two commands is only
the output in the pkg directory.

The old puppet module command will create a folder and a tarball with the module.

$ls pkg/
openstack-openstacklib-14.1.0/
openstack-openstacklib-14.1.0.tar.gz

While the PDK will only create the tarball in the
pkg folder. This should be no problem since I can't
find anywhere were we are using the generated folder
in the pkg folder.

This is part 2 in the series, it makes the actual change.

Depends-On: https://review.opendev.org/667589
Depends-On: https://review.opendev.org/667595
Needed-By: https://review.opendev.org/627554
Needed-By: https://review.opendev.org/627566
Change-Id: Id40ff665e92606455ef6b544f11394db2201c8b9