Move project config in-repo
This directory is only read by the "zuul" tenant in OpenDev Zuul,
so that it can contain the in-repo configuration for this project
without being seen by other users.

This is a copy of what is currently in project-config.  We will add
this, remove the settings from p-c, then begin moving other jobs
in-repo and updating this.

Change-Id: Ie41f2efcaace0ccaf63603a8a0df0de304910ca5