Shard project-pipeline config
We are about to grow a *lot* of jobs in this repo.  To make
the project-pipeline config manageable, shard it into a
each of the jobs files.  This will make it easy to add all the
jobs in the file to the project and avoid having a single giant
list of jobs.

Change-Id: I9677214ead6e5ce62153d64987bb54ebab2856e3