Support Rackspace in upload-logs-swift
Rackspace only provides unauthenticated access to object storage
via CDN in a non-openstack-standard way, so we need to do some
extra work to support that.

This also adds a helper script for testing which deletes a container
(since in order to do so, you may need to delete all the contents
first).

Some commented-out debug configuration lines are added for the
convenience of future developers.

Change-Id: I3d1fce824fb40136048f0988939d22f755236a59