Allow for overriding dns resolvers in install-kubernetes
By default kubernetes wants to use the local resolvers on the host
system. But in many cases the local resolver may be 127.0.0.1 which is a
different localhost on the host system then in the container network
namespaces. To address this disconnect you can supply a different
resolv.conf with a list of nameservers to use instead. Add support for
this.

Note that the test currently only uses ipv4 resolvers because by default
k8s has no support for ipv6.

Change-Id: I3c0d7d88613e2c9112e809f036aef3dfeaf7ab72