Switch PDF file name to doc-PROJECT.pdf
Further work on the PDF generation showed that many projects have
project logos as svg files with the name PROJECT.svg, those get
converted to PROJECT.pdf and collide with the default name used here.

Change the default to "doc-PROJECT.pdf" to have a unique name.

This is a change in the default but we only added this 11 days ago, so
I hope it's fine to change.

Change-Id: I4f2224633fd4edbc3121af53d0202796cf8a0a60