Flip the order of the emit-job-header tests
This more accurate reflects how this is run by opendev. Note this is not
strictly necessary to get in as the post-logs.yaml playbooks determine
where things are uploaded anyway. However, this may be slightly easier
for people to read in logs if things line up more closely to each other.

Depends-On: https://review.opendev.org/680476
Change-Id: Ied902ea2b8c7b965447e16f91400586ca1f708e6