Make opensuse-15 jobs non-voting
There is currently an issue with the mirror for this platform.

Jobs report

 http://<mirror>/opensuse/update/leap/15.1/oss/] Valid metadata not found at specified URL

We have checked the mirroring process for this directory, and it is
in-sync with the upstream mirror.  This means it must be related to
something from the actual source distribution.

Make this non-voting while we sort this out.

Note, a follow-on (If41da9089a961dc27cda954924c17104e622dacc) makes
the tools/update-test-platforms.py handle non-voting platforms.
However, there are additional problems with duplicate jobs that need
to be fixed too.  Thus this is a manual update.

Change-Id: I22d4cc02abaecb23d85aec3d415b43501ab1e0d2