update-test-platforms.py : handle non-voting jobs
Add a non-voting platform dictionary so we can easily unblock the gate
via regenerating the automated jobs if one platform is having
temporary issues.  This adds "-nv" to the job name to make it clear
what is happening, and sets the job to "voting: false".

opensuse-15 is in this list, as a follow-on from the manual
(gate-unblocking) update I22d4cc02abaecb23d85aec3d415b43501ab1e0d2

Also, if a non-autogenerated job defined itself as non-voting, it
should not be put in the gate queue.  Handle this too.

Note that regenerating at this point is a no-op as the changes are
manually introduced in the prior change
I22d4cc02abaecb23d85aec3d415b43501ab1e0d2.

Change-Id: If41da9089a961dc27cda954924c17104e622dacc