Test swift upload logs without encoding type for gzip files
When retrieving gzipped files out of swift with gzip encoding set some
swift implementations return a decompressed version of the file if your
client cannot accept gzip encoding. This causes problems when the file
you want is actually compressed like a .tar.gz. Instead we avoid setting
the encoding type of these files forcing swift to give them back as is.

This change will be followed by a change that updates the production
role.

Change-Id: Ica27cf070d50fcafd879b6f8ccc3964afa9bf5e6