Do not keep owner when pulling zuul-output
Log synchronization fails when trying to preserve uid and guid
for a user and or group that doesn't exist on the executor.

This makes it so the owner and group is not preserved from the remote.

Simplify operations, use a single task with a loop instead of three
separate tasks.

Change-Id: I8ae937ba07e240ee4394374405aa841ad882022a