install-docker: option to install docker-compose
The install-docker role currently takes care of setting up Docker
however it lacks the ability to install docker-compose at the moment
which is quite useful.

This adds a variable which defaults to false that allows you to use the
role to install docker-compose.  It was decided to be put within the
install-docker role because that role already manages the repository
configuration and therefore provide consistency with the repository that
is used to install docker-compose.

Change-Id: Ia6f4ed88b4baf53bd671c7a7b48a1564ff2c866d