Add pause-buildset-registry role
This is a simple utility that provides the zuul_return functionality
for repos that need it encapsulating the existence of zuul_return.

For instance, system-config wanted to start doing this in a playbook
but linting the playbooks became super hard because of the new need
to have zuul installed for the zuul_return library. Just keep it here
in the family.

Change-Id: I6035009455f6434c2d6cfd976c7c6f414bdc2ae7