install-javascript-packages: add tox_constraints_file
I can't exactly say why, but the horizon postinstall in package.json
runs "tox -e npm" [1].  This then installs from requirements [2].

The requirements gate run a cross-horizon-npm job that ends up calling
this.  The short story is that if you modify upper-constraints.txt in
a requirements change, it is not picked up by this job.

Add tox_constraints_file to this job, so we can set the path to the
zuul checked-out version for the test.

[1] https://opendev.org/openstack/horizon/src/commit/e515713084f0ce5d7874ffe4ef080fcc2e5760cb/package.json#L22
[2] https://opendev.org/openstack/horizon/src/commit/e515713084f0ce5d7874ffe4ef080fcc2e5760cb/tox.ini#L23

Needed-By: https://review.opendev.org/709389
Change-Id: Ia7dd5af19b42c2ab53dfb70f92e184b0dc215d03