install-kubernetes: install lz4 packages
minikube has added a new "feature" which pre-loads a blob of data
with everything it needs that we can't disable, with a fallback.

we're currently downloading ~500MB and failing to use it because
it's compressed by lz4 which we don't have in images, this change
adds that to the package list.

Change-Id: If74faf2b21d564fa23253ea2eb5378457f055735