Revert "upload-logs-swift: Create a download script"
This reverts commit acde44818d0827174b22290192b3c728c440a1c5 and
the testing part of b3f417a6e6fb9539d4f7aee46a123e7ced76d0b5.

We'd like to obtain more consensus on the download script before
we commit to this.  In particular, the new zuul manifest file may
make it possible to do this without adding the feature to the
log upload roles.

Change-Id: I959c44b4dac6cad6d1b3d82ba6bc0949c9c759ff