go-jobs: improve testing
The jobs currently are not running tests if we make changes to the
"go" role due to a typo in the files.  Also, the file name should
really be go-jobs.yaml because we're testing the jobs, not any specific
role.

This patch also adds support for verifying that it works across all
of the platforms.

Change-Id: Ie6c64936d01e368f8af789f6576f8425cfd9350c