fetch-zuul-cloner: use ensure-pip
Firstly, the "destination" variable was replaced with a virtualenv in
6bb1f6046b692ee2038eafd0965bd3095d6fa787; update the documentation and
prefix the name as is now standard for variables in this repo.  This
is not set anywhere so I think the prefixing is OK.

Call ensure-pip before installing, and use the virtualenv_command it
exports to create the environment for the zuul-cloner binary.

Story: #2007386
Task: #39311
Change-Id: I057191bf20ab1650a8926971dd386ce9cb1f03f5