ensure-pip: export ensure_pip_virtualenv_command
This makes ensure-pip export something that should be sensible for
users of pip: to put in their 'virtualenv_command' for their host.

The theory of operation is fairly simple; see if "python3 -m venv"
looks like it works, and set that as the command if it does.  If not,
set "virtualenv".  For sanity, we check if it works.

We pull in virtualenv in the Python 2 case, but for Python 3 we are
deliberately do not bring it as it is an unnecessary dependency.  If
jobs do require the actual `virtualenv` package, they should provision
it themselves ... except for Xenial, which, as described inline, has
issues.

Follow-on changes will convert existing zuul-jobs roles that install
tools into virtualenvs to use this argument.

Change-Id: Idad14c0e77eed5bf8df2c8f84f52fbdea2236a9f