tox: Don't inline python warnings
The sphinx regex is picking up python warnings.  We can disambiguate
the two because sphinx doesn't emit a space before the message but
the python warnings module does.  This adjusts the regex to do that
and adds a test file that includes warnings output.

If we need/want to make this more robust, we could also simply filter
the sphinx output for messages that start with " ?\w+Warning:".

We could also decide that we want to include python warnings.

Change-Id: Ib58cc1a2d4f673a0ce3d2a9cae306004559f5b52