Fix check_jobs_documented linter
If autoroles is used to document all roles - which is not the
case in this repo but in opendev/base-jobs, the linter would complain
about missing roles. Fix the linter to not complain.

This change is here to sync the two variants and avoid others that copy
it to run into the problem.

Change-Id: I25411157ff18858e4f4c121a6d3f0ba6ec793be0