Fail fetch-sphinx-tarball if no html exists
If no html is generated but a PDF, the role will succeed. This is wrong,
we always need HTML for the publish step where the logic is: Publish
HTML and optionally PDF.

I run into this when updating tox.ini and breaking html output, the jobs
passed but promote job failed.

Change-Id: I44d15da4c8e1ec0d1f7284daed69dbf5caa40bf0