fetch-sphinx-tarball: use remote_src true
Since ansible copies files as a tempfile to the remote host by default
setting remote_src: true when the files are on the same host gives better
error messages since otherwise the source path is not reflecting the real path.

See: http://paste.openstack.org/show/792400/
Where the file in "dest:" field of unarchive is copied to a tempfile
instead of just unarchiving the tarball directly.

Change-Id: Id7478a5a3171fdd2d3bc7531e66a09414e711d3a