Fix fetch-sphinx-tarball fails
I added previously a check that makes docs job fail in post if there's no
html output. This now leads to a POST_FAILURE when the normal build fails.

So, a FAILURE turned into POST_FAILURE, see
https://zuul.opendev.org/t/zuul/build/afbc8c39929a46039aa4d987971cdb4f .

Update the check in the fetch-sphinx-tarball role that it only fails
if the playbooks passed.

Change-Id: Iebd0bffcc5caaa6df97d432a712d83b64db8e780